Add files via upload
The Tic-Tac-Toe game in Java is like playing the classic pen-and-paper game on your computer screen. You have a grid of 3x3 squares, and you click on them to place either an 'X' or an 'O'. The game tells you whose turn it is and when someone wins or if it's a tie. It's a fun way to play the game without needing a pen and paper!
This is developed in JAVA with an interactive window which can be played in a player vs player mode.